+++ benchmarks/JOB-bench/src/datalevin_bench/core.clj
@@ -1,6 +1,7 @@
 (ns datalevin-bench.core
   (:require
    [datalevin.core :as d]
+   [clojure.java.io :as io]
    [clojure.string :as s]))
 
 (def schema
@@ -129,218 +130,314 @@
 (def role-type-base       50)
 (def title-base           80000000)
 
-(defn- add-comp-cast-type [datoms]
-  (reduce (fn [ds [id content]]
-            (conj! ds (d/datom (+ comp-cast-type-base (Long/parseLong id))
-                               :comp-cast-type/kind content)))
-          datoms (d/read-csv (slurp "data/comp_cast_type.csv"))))
+(defn- add-comp-cast-type []
+  (map (fn [[id content]]
+         (d/datom (+ comp-cast-type-base (Long/parseLong id))
+                  :comp-cast-type/kind content))
+       (d/read-csv (slurp "data/comp_cast_type.csv"))))
 
-(defn- add-company-type [datoms]
-  (reduce (fn [ds [id content]]
-            (conj! ds (d/datom (+ company-type-base (Long/parseLong id))
-                               :company-type/kind content)))
-          datoms (d/read-csv (slurp  "data/company_type.csv"))))
+(defn- add-company-type []
+  (map (fn [[id content]]
+         (d/datom (+ company-type-base (Long/parseLong id))
+                  :company-type/kind content))
+       (d/read-csv (slurp  "data/company_type.csv"))))
 
-(defn- add-kind-type [datoms]
-  (reduce (fn [ds [id content]]
-            (conj! ds (d/datom (+ kind-type-base (Long/parseLong id))
-                               :kind-type/kind content)))
-          datoms (d/read-csv (slurp "data/kind_type.csv"))))
+(defn- add-kind-type []
+  (map (fn [[id content]]
+         (d/datom (+ kind-type-base (Long/parseLong id))
+                  :kind-type/kind content))
+       (d/read-csv (slurp "data/kind_type.csv"))))
 
-(defn- add-link-type [datoms]
-  (reduce (fn [ds [id content]]
-            (conj! ds (d/datom (+ link-type-base (Long/parseLong id))
-                               :link-type/link content)))
-          datoms (d/read-csv (slurp "data/link_type.csv"))))
+(defn- add-link-type []
+  (map (fn [[id content]]
+         (d/datom (+ link-type-base (Long/parseLong id))
+                  :link-type/link content))
+       (d/read-csv (slurp "data/link_type.csv"))))
 
-(defn- add-role-type [datoms]
-  (reduce (fn [ds [id content]]
-            (conj! ds (d/datom (+ role-type-base (Long/parseLong id))
-                               :role-type/role content)))
-          datoms (d/read-csv (slurp "data/role_type.csv"))))
+(defn- add-role-type []
+  (map (fn [[id content]]
+         (d/datom (+ role-type-base (Long/parseLong id))
+                  :role-type/role content))
+       (d/read-csv (slurp "data/role_type.csv"))))
 
-(defn- add-info-type [datoms]
-  (reduce (fn [ds [id content]]
-            (conj! ds (d/datom (+ info-type-base (Long/parseLong id))
-                               :info-type/info content)))
-          datoms (d/read-csv (slurp "data/info_type.csv"))))
+(defn- add-info-type []
+  (map (fn [[id content]]
+         (d/datom (+ info-type-base (Long/parseLong id))
+                  :info-type/info content))
+       (d/read-csv (slurp "data/info_type.csv"))))
 
-(defn- add-movie-link [datoms]
-  (reduce
-    (fn [ds [id movie linked-movie link-type]]
+(defn- add-movie-link []
+  (mapcat
+    (fn [[id movie linked-movie link-type]]
       (let [eid (+ movie-link-base (Long/parseLong id))]
-        (-> ds
-            (conj! (d/datom eid :movie-link/movie
-                            (+ title-base (Long/parseLong movie))))
-            (conj! (d/datom eid :movie-link/linked-movie
-                            (+ title-base (Long/parseLong linked-movie))))
-            (conj! (d/datom eid :movie-link/link-type
-                            (+ link-type-base (Long/parseLong link-type)))))))
-    datoms (d/read-csv (slurp "data/movie_link.csv"))))
+        [(d/datom eid :movie-link/movie
+                  (+ title-base (Long/parseLong movie)))
+         (d/datom eid :movie-link/linked-movie
+                  (+ title-base (Long/parseLong linked-movie)))
+         (d/datom eid :movie-link/link-type
+                  (+ link-type-base (Long/parseLong link-type)))]))
+    (d/read-csv (slurp "data/movie_link.csv"))))
 
-(defn- add-aka-name [datoms]
-  (reduce
-    (fn [ds [id person name imdb-index name-pcode-cf
-            name-pcode-nf surname-pcode]]
+(defn- add-aka-name []
+  (mapcat
+    (fn [[id person name imdb-index name-pcode-cf
+         name-pcode-nf surname-pcode]]
       (let [eid (+ aka-name-base (Long/parseLong id))]
-        (cond-> (-> ds
-                    (conj! (d/datom eid :aka-name/person
-                                    (+ name-base (Long/parseLong person))))
-                    (conj! (d/datom eid :aka-name/name name)))
+        (cond-> [(d/datom eid :aka-name/person
+                          (+ name-base (Long/parseLong person)))
+                 (d/datom eid :aka-name/name name)]
           (not (s/blank? imdb-index))
-          (conj! (d/datom eid :aka-name/imdb-index imdb-index))
+          (conj (d/datom eid :aka-name/imdb-index imdb-index))
           (not (s/blank? name-pcode-cf))
-          (conj! (d/datom eid :aka-name/name-pcode-cf name-pcode-cf))
+          (conj (d/datom eid :aka-name/name-pcode-cf name-pcode-cf))
           (not (s/blank? name-pcode-nf))
-          (conj! (d/datom eid :aka-name/name-pcode-nf name-pcode-nf))
+          (conj (d/datom eid :aka-name/name-pcode-nf name-pcode-nf))
           (not (s/blank? surname-pcode))
-          (conj! (d/datom eid :aka-name/surname-pcode surname-pcode)))))
-    datoms (d/read-csv (slurp "data/aka_name.csv"))))
+          (conj (d/datom eid :aka-name/surname-pcode surname-pcode)))))
+    (d/read-csv (slurp "data/aka_name.csv"))))
 
-(defn- add-aka-title [datoms]
-  (reduce
-    (fn [ds [id movie title imdb-index kind production-year phonetic-code
-            episode-of season-nr episode-nr note]]
+(defn- add-aka-title []
+  (mapcat
+    (fn [[id movie title imdb-index kind production-year phonetic-code
+         episode-of season-nr episode-nr note]]
       (let [eid (+ aka-title-base (Long/parseLong id))]
-        (cond-> (-> ds
-                    (conj! (d/datom eid :aka-title/movie
-                                    (+ title-base (Long/parseLong movie))))
-                    (conj! (d/datom eid :aka-title/title title)))
+        (cond-> [(d/datom eid :aka-title/movie
+                          (+ title-base (Long/parseLong movie)))
+                 (d/datom eid :aka-title/title title)]
           (not (s/blank? imdb-index))
-          (conj! (d/datom eid :aka-title/imdb-index imdb-index))
+          (conj (d/datom eid :aka-title/imdb-index imdb-index))
           (not (s/blank? kind))
-          (conj! (d/datom eid :aka-title/kind
-                          (+ kind-type-base (Long/parseLong kind))))
+          (conj (d/datom eid :aka-title/kind
+                         (+ kind-type-base (Long/parseLong kind))))
           (not (s/blank? production-year))
-          (conj! (d/datom eid :aka-title/production-year
-                          (Long/parseLong production-year)))
+          (conj (d/datom eid :aka-title/production-year
+                         (Long/parseLong production-year)))
           (not (s/blank? phonetic-code))
-          (conj! (d/datom eid :aka-title/phonetic-code phonetic-code))
+          (conj (d/datom eid :aka-title/phonetic-code phonetic-code))
           (not (s/blank? episode-of))
-          (conj! (d/datom eid :aka-title/episode-of
-                          (+ title-base (Long/parseLong episode-of))))
+          (conj (d/datom eid :aka-title/episode-of
+                         (+ title-base (Long/parseLong episode-of))))
           (not (s/blank? season-nr))
-          (conj! (d/datom eid :aka-title/season-nr (Long/parseLong season-nr)))
+          (conj (d/datom eid :aka-title/season-nr (Long/parseLong season-nr)))
           (not (s/blank? episode-nr))
-          (conj! (d/datom eid :aka-title/episode-nr (Long/parseLong episode-nr)))
+          (conj (d/datom eid :aka-title/episode-nr (Long/parseLong episode-nr)))
           (not (s/blank? note))
-          (conj! (d/datom eid :aka-title/note note)))))
-    datoms (d/read-csv (slurp "data/aka_title.csv"))))
+          (conj (d/datom eid :aka-title/note note)))))
+    (d/read-csv (slurp "data/aka_title.csv"))))
 
-(defn- add-company-name [datoms]
-  (reduce
-    (fn [ds [id name country-code imdb-id name-pcode-nf name-pcode-sf]]
+(defn- add-company-name []
+  (mapcat
+    (fn [[id name country-code imdb-id name-pcode-nf name-pcode-sf]]
       (let [eid (+ company-name-base (Long/parseLong id))]
-        (cond-> (conj! ds (d/datom eid :company-name/name name))
+        (cond-> [(d/datom eid :company-name/name name)]
           (not (s/blank? country-code))
-          (conj! (d/datom eid :company-name/country-code country-code))
+          (conj (d/datom eid :company-name/country-code country-code))
           (not (s/blank? imdb-id))
-          (conj! (d/datom eid :company-name/imdb-id (Long/parseLong imdb-id)))
+          (conj (d/datom eid :company-name/imdb-id (Long/parseLong imdb-id)))
           (not (s/blank? name-pcode-nf))
-          (conj! (d/datom eid :company-name/name-pcode-nf name-pcode-nf))
+          (conj (d/datom eid :company-name/name-pcode-nf name-pcode-nf))
           (not (s/blank? name-pcode-sf))
-          (conj! (d/datom eid :company-name/name-pcode-sf name-pcode-sf)))))
-    datoms (d/read-csv (slurp "data/company_name.csv"))))
+          (conj (d/datom eid :company-name/name-pcode-sf name-pcode-sf)))))
+    (d/read-csv (slurp "data/company_name.csv"))))
 
-(defn- add-complete-cast [datoms]
-  (reduce
-    (fn [ds [id movie subject status]]
+(defn- add-complete-cast []
+  (mapcat
+    (fn [[id movie subject status]]
       (let [eid (+ complete-cast-base (Long/parseLong id))]
-        (-> ds
-            (conj! (d/datom eid :complete-cast/movie
-                            (+ title-base (Long/parseLong movie))))
-            (conj! (d/datom eid :complete-cast/subject
-                            (+ comp-cast-type-base (Long/parseLong subject))))
-            (conj! (d/datom eid :complete-cast/status
-                            (+ comp-cast-type-base (Long/parseLong status)))))))
-    datoms (d/read-csv (slurp "data/complete_cast.csv"))))
+        [(d/datom eid :complete-cast/movie
+                  (+ title-base (Long/parseLong movie)))
+         (d/datom eid :complete-cast/subject
+                  (+ comp-cast-type-base (Long/parseLong subject)))
+         (d/datom eid :complete-cast/status
+                  (+ comp-cast-type-base (Long/parseLong status)))]))
+    (d/read-csv (slurp "data/complete_cast.csv"))))
 
-(defn- add-keyword [datoms]
-  (reduce
-    (fn [ds [id keyword phonetic-code]]
+(defn- add-keyword []
+  (mapcat
+    (fn [[id keyword phonetic-code]]
       (let [eid (+ keyword-base (Long/parseLong id))]
-        (-> ds
-            (conj! (d/datom eid :keyword/keyword keyword))
-            (conj! (d/datom eid :keyword/phonetic-code phonetic-code)))))
-    datoms (d/read-csv (slurp "data/keyword.csv"))))
+        [(d/datom eid :keyword/keyword keyword)
+         (d/datom eid :keyword/phonetic-code phonetic-code)]))
+    (d/read-csv (slurp "data/keyword.csv"))))
 
-(defn- add-char-name [datoms]
-  (reduce
-    (fn [ds [id name imdb-index imdb-id name-pcode-nf surname-pcode]]
+(defn- add-char-name []
+  (mapcat
+    (fn [[id name imdb-index imdb-id name-pcode-nf surname-pcode]]
       (let [eid (+ char-name-base (Long/parseLong id))]
-        (cond-> (conj! ds (d/datom eid :char-name/name name))
+        (cond-> [(d/datom eid :char-name/name name)]
           (not (s/blank? imdb-index))
-          (conj! (d/datom eid :char-name/imdb-index imdb-index))
+          (conj (d/datom eid :char-name/imdb-index imdb-index))
           (not (s/blank? imdb-id))
-          (conj! (d/datom eid :char-name/imdb-id (Long/parseLong imdb-id)))
+          (conj (d/datom eid :char-name/imdb-id (Long/parseLong imdb-id)))
           (not (s/blank? name-pcode-nf))
-          (conj! (d/datom eid :char-name/name-pcode-nf name-pcode-nf))
+          (conj (d/datom eid :char-name/name-pcode-nf name-pcode-nf))
           (not (s/blank? surname-pcode))
-          (conj! (d/datom eid :char-name/surname-pcode surname-pcode)))))
-    datoms (d/read-csv (slurp "data/char_name.csv"))))
+          (conj (d/datom eid :char-name/surname-pcode surname-pcode)))))
+    (d/read-csv (slurp "data/char_name.csv"))))
 
-(persistent! (-> (transient [])
-                 ;; add-comp-cast-type
-                 ;; add-company-type
-                 ;; add-kind-type
-                 ;; add-link-type
-                 ;; add-role-type
-                 ;; add-info-type
-                 ;; add-movie-link
-                 ;; add-aka-name
-                 ;; add-aka-title
-                 ;; add-company-name
-                 ;; add-complete-cast
-                 ;; add-keyword
-                 add-char-name
-                 ))
+(defn- add-movie-companies []
+  (mapcat
+    (fn [[id movie company company-type note]]
+      (let [eid (+ movie-companies-base (Long/parseLong id))]
+        (cond-> [(d/datom eid :movie-companies/movie
+                          (+ title-base (Long/parseLong movie)))
+                 (d/datom eid :movie-companies/company
+                          (+ company-name-base
+                             (Long/parseLong company)))
+                 (d/datom eid :movie-companies/company-type
+                          (+ company-type-base
+                             (Long/parseLong company-type)))]
+          (not (s/blank? note))
+          (conj (d/datom eid :movie-companies/note note)))))
+    (d/read-csv (slurp "data/movie_companies.csv"))))
 
-(defn- add-movie-companies [ds]
-  )
+(defn- add-movie-info [reader]
+  (mapcat
+    (fn [[id movie info-type info note]]
+      (let [eid (+ movie-info-base (Long/parseLong id))]
+        (cond-> [(d/datom eid :movie-info/movie
+                          (+ title-base (Long/parseLong movie)))
+                 (d/datom eid :movie-info/info-type
+                          (+ info-type-base
+                             (Long/parseLong info-type)))
+                 (d/datom eid :movie-info/info info)]
+          (not (s/blank? note))
+          (conj (d/datom eid :movie-info/note note)))))
+    (d/read-csv reader)))
 
-(defn- add-movie-info [ds]
-  )
+(defn- add-movie-info-idx [reader]
+  (mapcat
+    (fn [[id movie info-type info note]]
+      (let [eid (+ movie-info-idx-base (Long/parseLong id))]
+        (cond-> [(d/datom eid :movie-info-idx/movie
+                          (+ title-base (Long/parseLong movie)))
+                 (d/datom eid :movie-info-idx/info-type
+                          (+ info-type-base (Long/parseLong info-type)))
+                 (d/datom eid :movie-info-idx/info info)]
+          (not (s/blank? note))
+          (conj (d/datom eid :movie-info-idx/note note)))))
+    (d/read-csv reader)))
 
-(defn- add-movie-info-idx [ds]
-  )
+(defn- add-movie-keyword [reader]
+  (mapcat
+    (fn [[id movie keyword]]
+      (let [eid (+ movie-keyword-base (Long/parseLong id))]
+        [(d/datom eid :movie-keyword/movie
+                  (+ title-base (Long/parseLong movie)))
+         (d/datom eid :movie-keyword/keyword
+                  (+ keyword-base (Long/parseLong keyword)))]))
+    (d/read-csv reader)))
 
-(defn- add-movie-keyword [ds]
-  )
+(defn- add-name [reader]
+  (mapcat
+    (fn [[id name imdb-index imdb-id gender name-pcode-cf name-pcode-nf
+         surname-pcode]]
+      (let [eid (+ name-base (Long/parseLong id))]
+        (cond-> [(d/datom eid :name/name name)]
+          (not (s/blank? imdb-index))
+          (conj (d/datom eid :name/imdb-index imdb-index))
+          (not (s/blank? imdb-id))
+          (conj (d/datom eid :name/imdb-id (Long/parseLong imdb-id)))
+          (not (s/blank? gender))
+          (conj (d/datom eid :name/gender gender))
+          (not (s/blank? name-pcode-cf))
+          (conj (d/datom eid :name/name-pcode-cf name-pcode-cf))
+          (not (s/blank? name-pcode-nf))
+          (conj (d/datom eid :name/name-pcode-nf name-pcode-nf))
+          (not (s/blank? surname-pcode))
+          (conj (d/datom eid :name/surname-pcode surname-pcode)))))
+    (d/read-csv reader)))
 
-(defn- add-name-base [ds]
-  )
+(defn- add-person-info [reader]
+  (mapcat
+    (fn [[id person info-type info note]]
+      (let [eid (+ person-info-base (Long/parseLong id))]
+        (cond-> [(d/datom eid :person-info/person
+                          (+ name-base (Long/parseLong person)))
+                 (d/datom eid :person-info/info-type
+                          (+ info-type-base (Long/parseLong info-type)))
+                 (d/datom eid :person-info/info info)]
+          (not (s/blank? note))
+          (conj (d/datom eid :person-info/note note)))))
+    (d/read-csv reader)))
 
-(defn- add-person-info [ds]
-  )
+(defn- add-title [reader]
+  (mapcat
+    (fn [[id title imdb-index kind production-year imdb-id phonetic-code
+         episode-of season-nr episode-nr series-years]]
+      (let [eid (+ title-base (Long/parseLong id))]
+        (cond-> [(d/datom eid :title/title title)
+                 (d/datom eid :title/kind
+                          (+ kind-type-base (Long/parseLong kind)))]
+          (not (s/blank? imdb-index))
+          (conj (d/datom eid :title/imdb-index imdb-index))
+          (not (s/blank? production-year))
+          (conj (d/datom eid :title/production-year
+                         (Long/parseLong production-year)))
+          (not (s/blank? imdb-id))
+          (conj (d/datom eid :title/imdb-id (Long/parseLong imdb-id)))
+          (not (s/blank? phonetic-code))
+          (conj (d/datom eid :title/phonetic-code phonetic-code))
+          (not (s/blank? episode-of))
+          (conj (d/datom eid :title/episode-of
+                         (+ title-base (Long/parseLong episode-of))))
+          (not (s/blank? season-nr))
+          (conj (d/datom eid :title/season-nr (Long/parseLong season-nr)))
+          (not (s/blank? episode-nr))
+          (conj (d/datom eid :title/episode-nr (Long/parseLong episode-nr)))
+          (not (s/blank? series-years))
+          (conj (d/datom eid :title/series-years series-years)))))
+    (d/read-csv reader)))
 
-(defn- add-title [ds]
-  )
+(defn- add-cast-info [reader]
+  (mapcat
+    (fn [[id person movie person-role note nr-order role]]
+      (let [eid (+ cast-info-base (Long/parseLong id))]
+        (cond-> [(d/datom eid :cast-info/person
+                          (+ name-base (Long/parseLong person)))
+                 (d/datom eid :cast-info/movie
+                          (+ title-base (Long/parseLong movie)))
+                 (d/datom eid :cast-info/role
+                          (+ role-type-base (Long/parseLong role)))]
+          (not (s/blank? person-role))
+          (conj (d/datom eid :cast-info/person-role
+                         (+ char-name-base (Long/parseLong person-role))))
+          (not (s/blank? note))
+          (conj (d/datom eid :cast-info/note note))
+          (not (s/blank? nr-order))
+          (conj (d/datom eid :cast-info/nr-order
+                         (Long/parseLong nr-order))))))
+    (d/read-csv reader)))
 
-(defn- add-cast-info [ds]
-  )
-
-#_(def datoms (-> (transient [])
-                  add-comp-cast-type
-                  add-company-type
-                  add-kind-type
-                  add-link-type
-                  add-role-type
-                  add-info-type
-                  add-movie-link
-                  add-aka-name
-                  add-aka-title
-                  add-company-name
-                  add-complete-cast
-                  add-keyword
-                  add-char-name
-                  add-movie-companies
-                  add-movie-info
-                  add-movie-info-idx
-                  add-movie-keyword
-                  add-name-base
-                  add-person-info
-                  add-title
-                  add-cast-info
-                  persistent!))
-
-;;(def db (d/init-db datoms "db" schema {:closed-schema? true}))
+(def db (with-open [movie-info-rdr     (io/reader "data/movie_info.csv")
+                    movie-info-idx-rdr (io/reader "data/movie_info_idx.csv")
+                    movie-keyword-rdr  (io/reader "data/movie_keyword.csv")
+                    name-rdr           (io/reader "data/name.csv")
+                    person-info-rdr    (io/reader "data/person_info.csv")
+                    title-rdr          (io/reader "data/title.csv")
+                    cast-info-rdr      (io/reader "data/cast_info.csv")]
+          (time (d/init-db
+                  (concat
+                    (add-comp-cast-type)
+                    (add-company-type)
+                    (add-kind-type)
+                    (add-link-type)
+                    (add-role-type)
+                    (add-info-type)
+                    (add-movie-link)
+                    (add-aka-name)
+                    (add-aka-title)
+                    (add-company-name)
+                    (add-complete-cast)
+                    (add-keyword)
+                    (add-char-name)
+                    (add-movie-companies)
+                    (add-movie-info movie-info-rdr)
+                    (add-movie-info-idx movie-info-idx-rdr)
+                    (add-movie-keyword movie-keyword-rdr)
+                    (add-name name-rdr)
+                    (add-person-info person-info-rdr)
+                    (add-title title-rdr)
+                    (add-cast-info cast-info-rdr))
+                  "db" schema {:closed-schema? true
+                               :kv-opts        {:mapsize 20000}}))))